backoffice/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
django-admin-interface==0.26.0
django-admin-sortable2==2.1.10
django-admin-tailwind==0.0.31
django-baton==2.8.1
django-ckeditor==6.7.0
django-colorfield==0.10.1
django-js-asset==2.1.0
django-modeltranslation==0.18.11
idna==3.4
Pillow==10.0.1
PyJWT==2.8.0
python-slugify==8.0.1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
django-recaptcha==4.0.0
django-redis
pyotp
qrcode
passlib
